+++ src/app/components/footer/Footer.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+
+function Footer() {
+  return (
+    <footer className='flex w-full shrink-0 items-center justify-center bg-black text-xs text-neutral-400'>
+      <div className='flex w-full max-w-screen-xl justify-between px-12 py-4 lg:px-24'>
+        <p>✌️</p>
+        <p>2025 Red Oral </p>
+      </div>
+    </footer>
+  );
+}
+
+export default Footer;
