+++ src/app/components/nav/desktop/index.ts
@@ -1,0 +1,1 @@
+export { default } from './DesktopNav';
