+++ src/app/components/nav/mobile/index.ts
@@ -1,0 +1,1 @@
+export { default } from './MobileNav';
