+++ src/app/components/page-layout/index.ts
@@ -1,0 +1,1 @@
+export { default } from './PageLayout';
