+++ src/app/components/project/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Project';
