+++ src/app/layout.tsx
@@ -1,36 +1,28 @@
+import React from 'react';
 import type { Metadata } from 'next';
 import { Sen } from 'next/font/google';
 import NextTopLoader from 'nextjs-toploader';
+import PageLayout from '@/app/components/page-layout/PageLayout';
+import Footer from '@/app/components/footer';
 import './globals.css';
-import PageLayout from './components/PageLayout';
 
 const inter = Sen({ subsets: ['latin'] });
 
 export const metadata: Metadata = {
-  title: 'Red Oral - Full Stack Software Engineer',
+  title: 'Red Oral',
   description:
     'Full Stack Software Engineer with a multitude of technical skills. From mobile & web apps to REST APIs & serverless backends, I am experienced in a magnitude of development tools and processes.',
 };
 
 export default function RootLayout({ children }: { readonly children: React.ReactNode }) {
   return (
-    <html lang='en' className={`h-full scroll-smooth`}>
+    <html lang='en' className='h-full'>
       <body
         className={`${inter.className} dark flex h-full flex-col bg-neutral-100 dark:bg-neutral-950 dark:text-white`}
       >
-        <header></header>
-        <main className='flex-auto'>
-          <NextTopLoader color='#ff6961' />
-          <div className='flex flex-col'>
-            <PageLayout>{children}</PageLayout>
-          </div>
-        </main>
-        <footer className='flex w-full shrink-0 items-center justify-center bg-black text-xs text-neutral-400'>
-          <div className='flex w-full max-w-screen-xl justify-between px-12 py-4 lg:px-24'>
-            <p>✌️</p>
-            <p>2025 Red Oral </p>
-          </div>
-        </footer>
+        <NextTopLoader color='#ff6961' />
+        <PageLayout>{children}</PageLayout>
+        <Footer />
       </body>
     </html>
   );
